backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic-settings==2.4.0
pillow==10.4.0
pytesseract==0.3.10
google-generativeai==0.5.2
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.8
sqlalchemy==2.0.36
aiosqlite==0.20.0
